Request

+++ src/main/java/world/ucode/model/db/dao/DAOlot.java
@@ -80,12 +80,11 @@ public class DAOlot implements DAO<Lot, Integer>{
     }
 
 
-    public List<Lot> getAllLotsbyCategoris(List<String> categories, String tittle, long startPrice, int rate, int status, String login, int userid) {
+    public List<Lot> getAllLotsbyCategoris(List<String> categories, String tittle, double startPrice, int rate, int status, String login, int userid) {
         Transaction transaction = null;
         List <Lot> listOfLot = null;
         try (Session session = HibernateUtil.getSessionFactory().openSession()) {
             transaction = session.beginTransaction();
-
 
             Query query = Querystring.stringmaker(categories, session, tittle, startPrice, rate, status, login, userid);
             listOfLot = query.getResultList();
